# openapi-gpt.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sismed Marketing API",
    "version": "3.7.0",
    "description": "API para prospectos, imágenes de marketing y publicaciones de Instagram de Sismed. Nunca debe almacenar datos clínicos ni datos de pacientes.\n"
  },
  "servers": [
    {
      "url": "https://sismed-marketing-action.onrender.com"
    }
  ],
  "paths": {
...
    "/api/medios/{tipo}/{id}/eliminacion": {
      "get": {
        "operationId": "consultarEliminacionMedio",
        "summary": "Antes de eliminar: consultar archivo, tamaño, referencias, bloqueos y revision_borrado. Solo lectura.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "tipo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "imagen",
                "reel"
              ]
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vista previa de eliminación; mostrar tipo, ID, enlace, usos y bytes a Cristian."
          },
          "404": {
            "description": "Medio inexistente"
          }
        }
      }
    },
    "/api/medios/{tipo}/{id}": {
      "delete": {
        "operationId": "eliminarMedioMarketing",
        "summary": "Eliminar bytes de una imagen o reel con autorización explícita de Cristian. Conserva IDs e historial, no borra en Instagram. Nunca cancelar ni borrar dependencias automáticamente. Requiere consulta previa vigente.",
        "x-openai-isConsequential": true,
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "tipo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "imagen",
                "reel"
              ]
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "confirmacion",
                  "revision_borrado"
                ],
                "properties": {
                  "confirmacion": {
                    "type": "boolean",
                    "enum": [
                      true
                    ]
                  },
                  "revision_borrado": {
                    "type": "string",
                    "pattern": "^[a-f0-9]{64}$"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Archivo eliminado o previamente eliminado; bytes liberados e historial conservado."
          },
          "400": {
            "description": "Falta confirmación o parámetros inválidos"
          },
          "409": {
            "description": "Medio en uso o consulta obsoleta. No se realizó la eliminación."
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  }
}